--- app/core/clients/__init__.py
@@ -1,7 +1,0 @@
-from .rabbitmq import RabbitMQClient
-from .redis import RedisClient
-
-__all__ = [
-    "RabbitMQClient",
-    "RedisClient",
-]
