+++ app/services/v1/auth/__init__.py
@@ -1,4 +1,5 @@
-from .service import AuthService, AuthDataManager
+from .service import AuthService
+from .data_manager import AuthDataManager
 
 
 __all__ = ["AuthService", "AuthDataManager"]
